+++ pyperf/tests/test_examples.py
@@ -1,9 +1,9 @@
 import glob
 import os.path
 import sys
+import unittest
 
 from pyperf import tests
-from pyperf.tests import unittest
 
 
 ROOT_DIR = os.path.realpath(os.path.join(os.path.dirname(__file__), '..', '..'))