+++ pyperf/tests/test_timeit.py
@@ -5,11 +5,11 @@ import shutil
 import sys
 import tempfile
 import textwrap
+import unittest
 
 import pyperf
 from pyperf import tests
 from pyperf._timeit import Timer
-from pyperf.tests import unittest
 
 
 PERF_TIMEIT = (sys.executable, '-m', 'pyperf', 'timeit')